Add --parse-only
