+++ fpga_arch_viewer/tests/fixtures/invalid_arch.xml
@@ -1,0 +1,3 @@
+<architecture>
+  <unknown_top_level_tag/>
+</architecture>
